edits to the reviewer's comments have been added
