+++ package.json
@@ -43,7 +43,6 @@
     "classnames": "^2.3.1",
     "algoliasearch": "^4.13.0",
     "date-fns": "^2.27.0",
-    "docsearch.js": "^2.6.3",
     "next": "12.1.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
