+++ utils/debounced.ts
@@ -1,0 +1,18 @@
+function debouncePromise(fn, time: number) {
+  let timerId = undefined;
+
+  return function debounced(...args) {
+    if (timerId) {
+      clearTimeout(timerId);
+    }
+
+    return new Promise((resolve) => {
+      timerId = setTimeout(() => resolve(fn(...args)), time);
+    });
+  };
+}
+
+export const debounced = debouncePromise(
+  (items) => Promise.resolve(items),
+  500
+);
